requirements.txt
altair==5.2.0
argcomplete==1.10.3
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.8.2
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.1
compressed-rtf==1.0.6
debugpy==1.8.1
decorator==5.1.1
docx2txt==0.8
ebcdic==1.1.1
exceptiongroup==1.2.0
executing==2.0.1
extract-msg==0.28.7
gitdb==4.0.11
GitPython==3.1.42
idna==3.6
IMAPClient==2.1.0
importlib-metadata==7.0.1
ipykernel==6.29.2
ipython==8.21.0
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
linkedin-api @ git+https://github.com/tomquirk/linkedin-api.git@4d69958d1e31f9e5d962ef07712162e2dedaf440
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mdurl==0.1.2
nest-asyncio==1.6.0
numpy==1.26.4
olefile==0.47
packaging==23.2
pandas==2.2.0
parso==0.8.3
pdfminer.six==20191110
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pycryptodome==3.20.0
pydeck==0.8.1b0
Pygments==2.17.2
PyPDF2==3.0.1
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.1
python-pptx==0.6.23
pytz==2024.1
pyzmq==25.1.2
referencing==0.33.0
requests==2.31.0
rich==13.7.0
rpds-py==0.18.0
six==1.12.0
smmap==5.0.1
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.8.1
stack-data==0.6.3
streamlit==1.31.1
tenacity==8.2.3
textract==1.6.5
toml==0.10.2
toolz==0.12.1
tornado==6.4
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.0
validators==0.22.0
watchdog==4.0.0
wcwidth==0.2.13
xlrd==1.2.0
XlsxWriter==3.1.9
zipp==3.17.0
